+++ NEWS.md
@@ -1,3 +1,3 @@
 # raiden 0.1.0
+* Create `raidenApp()` and `gather_*()` functions
 
-* Initial CRAN submission.